Added file object to '_closable_objects' to close at end of request
This commit fixes ResourceWarnings in Python 3 where file objects
do not appear to be closed at the end of the request, and the
Django server will emit ResourceWarnings about unclosed files.
The `_closable_objects` list is managed by the base FileResponse
class in Django and will call `.close()` on objects at the end
of the request.